[ResultDB] Getting variantsWithUnexpectedResults in query if there are too many of them
If there are too many variants with unexpected results in an invocation, it's possible that the query params exceed the request payload limit(10MiB).

This CL will fix the issue by discarding the testVariants in params and running the same query again to get the variants and join with TestResults.

Bug: 1147631
Change-Id: I926e662a1883d84899cdc0d648ef5741b5bc416a