[filegraph] Implement git.Updater.
git.Updater is the heart of filegraph project.
It constructs the graph based on the git history, and the the way it
does this is the most interesting part.
It is also the most controversial / arbitrary part and is subject to
further improvements, based on evaluation results.

Bug: 1136280
Change-Id: I9cd715df901d7a6d2117cb49280548327ce8e0cd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2524072
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>