[filegraph] Add git.Load() function.
loadUpdatedGraph() function is generic enough to be also needed
in RTS code. Move it from the cli package to git package, and
call it Load().

Move common git functions package internal/gitutil.

New/changed code is primarily in load.go.
The code there is very similar to previous version, but also
refactored.
Code in other files is largely the same.

Bug: 1136280
Change-Id: I15b1232c7a8b70c6756c9f1a1d8e3594c89d30c4