Fix a bug where the vpython virtualenv lock was not passed to python.
This could conceivably cause the virtualenv to be deleted out from
under a running python instance by a concurrent vpython process.

On Mac, also remove the logic to dup the lock to fd3. This may
unintentionally close an existing fd3 that should be passed as-is
to the exec()d python. Instead, always just clear the close-on-exec
flag on the existing lock fd.

Change-Id: I3a99273b438c4052e0eb2168c83d45ac40c02ad9