[resultui] Restyle the blamelist tab.
This includes making the headings more tabular, and reducing some visual noise in the expanded entry to hopefully make it easier to read.

This also makes the blamelist entries collapsed by default. I think this should be fine for people who like to Ctrl+F now that we have the button to expand all.

Bug: 989245
Change-Id: I43f3ae6de7582701543ac0300ac21eb94ff13760
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2540126
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Joey Scarr <jsca@chromium.org>