[resultdb] allow clocks to be out of sync for at most 1 min
It has been reported that Recorder rejected test results because
`start_time + duration` is in the future.

It can happen because the clocks used in the bot and server are out of
sync, and Recorder should allow machine clocks to be out of sync for
a certain amount of time.

This CL adds 1 minute to the current time to avoid the problem.

R=chanli@chromium.org,nodir@chromium.org
CC=maruel@chromium.org

Bug: 1150844
Change-Id: I67b5bfa5b612f9b1ca90d361758c5484533f6d3b