[resultdb] allow clocks to be out of sync for at most 10 mins
It has been reported that Recorder rejected test results because
`start_time + duration` was in the future.

It can happen if the clocks used in the bot and server are out of sync.
Recorder should allow machine clocks to be out of sync for a certain
amount of time.

This CL adds 10 mins to the current time to avoid the problem.

R=chanli@chromium.org,nodir@chromium.org
CC=maruel@chromium.org

Bug: 1150844
Change-Id: I67b5bfa5b612f9b1ca90d361758c5484533f6d3b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2549432
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>