[BuildPage] Split summary header and content better
We are <br/> to separate header and content of summaryMarkdown. It may work for some cases, but for cases like '<div>header<br/></div>content'. In this case, header will be '<div>header' and content will be '</div>content'. As summaryMarkdown may potentially be anything, there will be chance that we will not be able to cleanly separate them.


In that case, one workaround may be just consider everything as content, if the end result of the breaking is not valid html.
Bug:1150140
Change-Id: If38019c7a46847cfd05e42bc6a9b0ddc46445d56