[BuildPage] Split summary header and content better
We are using <br/> to separate header and content of summaryMarkdown.

It may work for some cases, but for cases like '<div>header<br/></div>content', header will be '<div>header' and content will be '</div>content'.

As summaryMarkdown may potentially be anything, there will be chance that we will not be able to cleanly separate them. In that case, one workaround may be just consider everything as content, if the end result of the breaking is not valid html.

For this CL, if <br/> tag is enclosed in another tag, we consider summary cannot be cleanly separated, and everything should be treated as content. In this case, everything will be rendered in summaryHtml, just like legacy build page.

Also, if header contains <li> tag, we consider it is not valid as well. There maybe other tags, but we can add in later.

Bug:1150140
Change-Id: If38019c7a46847cfd05e42bc6a9b0ddc46445d56
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2552304
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>