[filegraph] Add family distance
Files in the same directory tend to be related.
Derive relatedness from the file structure by establishing the minimum
distance between parents and children. Report the parent-child
relatinship as an edge with a configured const distance.
It must be on the same "scale" as the distances derived from git,
so a default value is provided. It was computed based on experimens
in chromium.

Bug: 1136280
Change-Id: I2fe14a7c91da8a5af08757286b71ee28378fac5a