[rts] Move file list fetching to presubmit-history
Currently the evaluation framework is responsible for fetching the list
of changed files for a given patchset. This complicates its code
and is not actually its job to do. It would be simpler if the historical
files were self-sufficient.

Eventually the list of files will be fetched from BigQuery, but
currently there isn't enough BigQuery data with populated changed files.
In the meantime, move gerrit data fetching from evaluation framework
to presubmit-history subcommand, which runs BigQuery queries.

Keep the code simple: rejections.go does not try to parallelize Gerrit
fetch. Doing so would be complicated because the order of fetched
BigQuery rows and rejection fragments is significant. This is not too
bad because most of fetched rows are test failures of the same patchset,
e.g. there is only ~1500 failed patchsets in October.

Bug: 1112125
Change-Id: Ib7db0f3bbcf5c655a912ea558025b2c6ff36de26