[rts] Add distance concept to evaluation
Currently the process of tuning the max-distance is manual: the
developer is looking for a threshold that produces an acceptable
change recall, and then evaluates efficiency with that threshold.

To automate the search for the optimal max-distance, the first step is
to generalize the signature of the RTS algorithm by changing its output
from [set of skipped tests] to [distance for each test].
This way, the algorithm does not make a decision to skip a test, but
only provides a distance. Then the evaluation framework makes a decision
based on max-distance.

The next step will be finding the minimal distance threshold that
satisfies minimal ChangeRecall and TestRecall.

R=tandrii@chromium.org

Bug: 1112125
Change-Id: I952cfb336897f44909481af7513bd56d2036561e