[rts] Introduce test ranking
Let the RTS algorithm to rank the tests, where rank is one-based index
in the list of all existing tests, sorted by distance.

Add -min-run flag which is the minimum number of ranked tests to run.

This technique alone, with -min-run 10k and without -max-distance,
achieves 99.93% ChangeRecall and 51.53% savings
on 1w-worth data in November, over 2.6k rejections and 45h of testing.

Bug: 1112125
Change-Id: Ic415c6a34ef03a24cf81f4728410caf1aa644375