[server/tq] add support tqtesting.StopAfterTask.
This isn't perfect if combined with ParallelExecute,
because there is an unavoidable window between task handler
returning and StopAfterTask evaluation being done,
during which arbitrary number of tasks can be executed.
On the other hand, this is much closer to reality.

For testing tasks which may recurse unbounded number of times,
I expect that the serial execution with StopAfterTask will
suffice to check most of the functional correctness.

For larger integration test, test will be terminated by cancelling
context passed to Scheduler.Run based on ether passage of
simulated time or from TaskFailed/TaskSucceeded callback.

R=vadimsh

Change-Id: I09cd033f316435b19bb09ff81b012fb5e5f32348