cv: implement gerritfake.ListChanges.
It'd have been trivial, if not for the need to interpret the query
string, which takes 160+ LOC.

In theory, AIP-160 filter spec https://google.aip.dev/160 should be a
reasonable fit, though I didn't carefully verify that Gerrit grammar is
exact subset.  However, I couldn't find any Go library that implements
parser for the spec after 40 minutes of searching :(

So, I rolled naive tokenizer and a giant switch to recognize predicates.
It's not perfect, but catches typical mistakes and should be good enough
to start. In the future, I hope we can re-use a proper library.

[1] https://google.aip.dev/160

R=qyearsley, yiwzhang

Bug: 1150185
Change-Id: Ic8d130b4a35eae2cea6e2d84afa3985d533429d3