cv: implement ProjectManager mailbox structure.
This CL is ~65% boilerplate, but contains working pipeline
of sending messages to and receiving them by project manager.

I thought of a making it a library. In theory it's doable, even without
generics:
 * Payload proto message (see internal/events.proto). This could be
   re-used, but we'd need to stick both Run and Project events together.
 * Task type proto messages (internal/tasks.proto). These could be
   re-used by specifying RunManager or ProjectManager as a target,
   but then we'd need to share 1 TQ queue, which isn't good.
 * Datastore kind of event entity. Strictly speaking could be re-used,
   but this would be confusing, so ideally we should set kind on the fly,
   while re-using the struct. This is actually possible using datastore
   MetaGetter interface.

Another way is to generate the boilerplaty code from scratch.
But since we only need this one more time, I think it's easier to
copy-paste this by hand to RunManager.

R=qyearsley, yiwzhang

Bug: 1156241
Change-Id: I392e02a272a70cf2dc76fc0113f990ff80158322
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2575664
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>