[buildbucket] Update documented behavior of Batch rpc.
Previously Batch was meant to run all 'write' operations before all
'read' operations. However, in the vast majority of cases this should
not matter:
  * Get operations would require guessing the ID of a build you just
    scheduled; clearly not supported.
  * Search operations are mostly at the whimsy of the Datastore indexes;
    no way to guarantee that a scheduled build would be reflected in a
    query.

The one notable exception would be scheduling a Build in a BuildSet which
you then query; because BuildSet queries are special, they are actually
consistent, and you could reliably see the build you just scheduled show
up in the result set. However it's also easy to just do that on the client
side when you get the response back: add the scheduled build to the query
result.

The guarantee that write ops happen before read ops means that our ability
to transition the Batch RPC from Python to Go without a potentially large
increase in latency is quite limited. Removing this promise will allow us
to do read and write operations in parallel, potentially on both backends.

R=ddoman, smut, tandrii

Change-Id: Ia83568fc586770f1952c0c7c4ce6f94bf9f7b0b0