[ResultDB] UI-specific RPC: query first page of test variants with only expected results.
This is a separate query from the one for test variants with unexpected results.

Also in this change, add page token. It contains test variant type, test id and variant hash. And the test variant type in the token can help to  determine which query to run.

Bug: 1112127
Change-Id: I284569927bc4b23305f6ce1eb6ad69c01ce7300c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2575982
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>