[rts] Introduce test ranking
Let the RTS algorithm rank the tests, where rank is one-based index
in the list of all existing tests, sorted by distance.

Add -max-rank flag: all tests with rank <=maxRank are selected,
regardless of -max-distance.

This technique alone, with -max-rank 10k and without -max-distance,
achieves 99.93% ChangeRecall and 51.53% savings
on 1w-worth data in November, over 2.6k rejections and 45h of testing.

Ideally we find a perfect combination of -max-run and -max-distance
which achieves the best scores.

R=tandrii@chromium.org

Bug: 1112125
Change-Id: Ic415c6a34ef03a24cf81f4728410caf1aa644375