[resultdb] Fix flaky timeout in history.Query.collect()
This is caused because `wi.resultsC` is not guaranteed to be closed
until after a worker has picked up `wi`, but this may not happen if all
workers have already terminated (e.g. due to the context expiring).

So, if the collector has already blocked on the results channel of a
workItem that is _not_ going to be picked up, it will wait forever.

This race is fixed by replacing the range over `wi.resultsC` with a
loop over a select that checks the context and receives on the channel
simultaneously.

R=chanli,nodir,ddoman

Change-Id: I82b150373406ce3ff1c83970dcb0f9b2028cdb6a