[rts] Compute thresholds
Currently the user has to manually supply -max-distance and -max-rank
values and hope that it achieves the target ChangeRecall.
This can be automated because functions ChangeRecall=f(distanceThreshold) and
ChangeRecall=g(rankThresholds) are monotonically increasing, so it is
possible to compute the thresholds based on the desired ChangeRecall.

The ChangeRecall is not set in stone either: the user might choose a
ChangeRecall based on savings as well. So having to fix ChangeRecall
is not great either. Ideally the user sees a plot Savings vs ChangeRecall.
Note that the function Savings=h(ChangeRecall) is monotonic too.

All of this is possible to compute in O(historical_records) time
complexity thanks to the monotonicity of all these functions.
Specifically, we don't have to rerun the algorithm for each possible
ChangeRecall.

The key insight is that given
1) a target integer ChangeRecall N%
2) the list of Affectedness of the "closest test" for each change
   rejection in the data set.
We can compute the minimal distance threshold simply by taking the
N-th percentile of the distances. This means we can simply sort
all Affectedness values and compute percentiles => we have distance
thresholds for each possible integer ChangeRecall score.
Same process for ranks. Rank and distance distributions are computed
independently from each other.

The remaining problems are
- we need to somehow combine two thresholds
- prevent overfitting
- compute savings.
This can be solved as follows:
1) Build a 100x100 grid from the two 100-len arrays we acquired
from the "training set" (above)
2) For each change rejection in a "eval set", compute the Affectedness
of the "closest test" and incorporate it into the grid,
   e.g. for grid row, increment a counter in the matching rank bucket.
3) Similarly for test durations.

The output of this process is a essentially 100*100 rows with info:
- Distance threshold and the ChangeRecall achieved by this
  threshold alone on the
- Rank threshold and ChangeRecall achieved by this threshold alone on
  the training set
- ChangeRecall/TestRecall/Savings achieved by the combination of the thresholds on
  the eval set.
This solves all three problems:
- we combine the two thresholds by picking the combination with the best
  Savings and acceptable ChangeRecall
- we prevent overfitting by having separate training and eval sets,
  and bailing if ChangeRecall/Savings achieved on two sets differ
  significantly.
- We compute savings.

This looks like classic "grid search" in ML applications where
the ML practitioner searches for the best combination of
hyper-parameters by training the model for each possible combination.
The big difference is that typically grid-search is
O(number-of-hyper-parameter-combinations * data-points), while here
we do only one data pass and put each data point to the grid.

This is the first CL in a series that implements this.
This CL splits -history flag into -training-set and -eval-set,
and computes the grid, but does not yet use it.

Bug: 1112125
Change-Id: Idd8feab5dade93f7ae9af99b88e0a0a5bca94a84