[rts] Switch to the threshold grid
This is a part of
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2588088
When processing rejections and durations, do not decide whether
the rejection is preserved or a test is executed, but instead increment
counters in the grid.

Drop -max-distance and -max-rank flags.

Drop periodic updates for now, but print results in the end.
They were important to stop evaluation early when it was obvious
that the manually provided threshold is bad. We don't have that
problem anymore.

Bug: 1112125
Change-Id: I77d5246a8e40b6dd6da1620b6f15971d213c70fb