[led] Make edit-isolted be compatible with RBE-CAS
For the command `led edit-isolated`, the previous flow is to download
from isolate to local tmp folder, make changes in tmp and upload this
folder to isolate. Now, the new flow is to download from isolate or
rbe-cas, make changes and upload this folder to rbe-cas.

Bug:1145959
Change-Id: I342b27cb0c03bc3e1f2fc96702e8df5621953934
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2559652
Commit-Queue: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>