[rts] Switch to the threshold grid
This is the 2nd out of two CLs fully described in
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2588088

When processing rejections and durations, do not try to decide whether
to execute a test or whether a rejection got preserved. Instead increment
appropriate counters in the grid. After processing all data,
print a table depicting Savings=f(ChangeRecall) function.

Drop -max-distance and -max-rank flags.

Drop periodic updates for now, but print results in the end.
Periodic updates were important to stop evaluation early on when it was
obvious that the manually provided threshold is bad. We don't have that
problem anymore.

Bug: 1112125
Change-Id: I77d5246a8e40b6dd6da1620b6f15971d213c70fb