[resultdb] Fix queries with page token
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2578738
broke QueryTestResults for an unexpected results, with a page token.
It means if all unexpected results do not fit the first page, Milo
will fail to load all unexpected results.

The problem is that WHERE clause is missing, but `AND <condition>` is
appended, i.e. there is `AND` without `WHERE`.
Add `WHERE true`.

Change-Id: Ic87da25777bdf515721dc4c30238b0b84f9fdac4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2598259
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>