[rts] Switch to the threshold grid
This is the 2nd out of two CLs fully described in
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2588088

There is a minor departure.

The second history file is not used
for safety evaluation, but only for efficiency.
We have much more test durations than rejections. Normally
the processed rejection history file contains all rejections in a
month or more, where duration data includes 1% of one day or one week.
The second file covers a too small period to have enough rejections.

The split into training and eval sets is important in ML applications
because their model is derived from the training set. In our case,
the model is not. Only thresholds are chosen based on data.
The need for training|eval set split is not as obvious as in
traditional ML applications. Regardless, we can add it later.

Rename -training-set flag to -rejections. Use the same file to
first choose the thresholds, and then evaluate each threshold
combinations.

Rename -eval-set to -durations. Use the file to evaluate efficiency.

After processing all data,
print a table depicting Savings=f(ChangeRecall) function.

Drop -max-distance and -max-rank flags.

Drop periodic updates for now, but print results in the end.
Periodic updates were important to stop evaluation early on when it was
obvious that the manually provided threshold is bad. We don't have that
problem anymore.

Overall, evaluation code is now ~300 lines of code.

Bug: 1112125
Change-Id: I77d5246a8e40b6dd6da1620b6f15971d213c70fb