[rts] Move code around.
Struct Affectedness is not specific to evaluation.
It is needed "in production" too.
Move it to rts package.

thresholdGrid is more related to evaluation (the only usage)
than to the result. Move from result.go to eval.go.
Move quantiles() and closest() functions to eval.go too.

No "real" code changes.

Bug: 1090668
Change-Id: I6e81601d92eae18f79a88b69a54c469e1a10567e