[luciexe/build] Remove LogFile until we have a mockable filesystem design.
I expected that we'd be able to use the upcoming "io/fs" to build a
mockable filesystem standard for luciexe programs. However it appears that
the "io/fs" proposal is for read-only filesystems.

We can restore this once we have a proposed solution for a vfs system.

R=yiwzhang

Bug: 1157581
Change-Id: I554014c2f0e34fc8ff0bbd97c199f41af1173fa6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2614272
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>