[server] Integrate with server/warmup package.
The token server uses it to warm up some of its caches. It is not
essential, but nice to have.

One complication is that tsmon package warmup assumed GAE server
environment. Add a hack to recognize luci/server environment.
Hopefully it will simplify once we don't need to support multiple
environments.

R=tandrii@chromium.org, iannucci@chromium.org
BUG=1165454

Change-Id: Ia724dc5b9bda6726fa9a5bba971ddc225d550cad