[prpc] Make it deployable to GAE.
http.Header.Values was introduced in go1.14 and is not available
on GAE. It appears the only difference from accessing the map
directly is canonicalization of the key. But "Accept-Encoding" is
already in the canonical form, so we can just look in the map
directly.

R=nodir@chromium.org
BUG=1166257

Change-Id: I9e0df9ab2f447852f3dd9fc202caf8c95cf28a0e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2631688
Auto-Submit: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>