[common/testing] Update ShouldErrLike to accept multiple arguments.
This allows patterns like:

   So(fn(), ShouldErrLike,
     "a substring",
     "another substring")

Which is good to allow tests to match multiple keywords in the error
without having an overly brittle test that matches a super-long and
super-specific error message. This becomes even more compelling with
ShouldPanicLike, which relies on ShouldErrLike for its heavy lifting.

This also removes the strange formulation (which I don't think
I've ever seen in the tests):

  So(err, ShouldErrLike)

Which is a worse way to say

  So(err, ShouldBeNil)

R=yiwzhang

Bug: 1157581
Change-Id: I21c0a356bab27504081c3d7629b633d89a2a1e9e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2641328
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>