[streamclient] Improve Fake API.
This fixes a long-running inconsistency; Programs dealing with streamclient
invariably obtain that client via the `bootstrap` package, which is oriented
around extracting the client from the environment.

This new API allows temporary in-memory registration of Fakes to allow
`streamclient.New` to work with a simple string argument pulled from, say,
an environment variable.

This introduces the ability to have multiple Clients connected to the same
Fake at the same time, perhaps with different namespaces (functionality I
suspect I'll need to properly test the luciexe/build library).

R=yiwzhang

Bug: 1157581
Change-Id: I52f63f342e5ae2b16b2258721346ab7b4fcc9acb