[buildbucket] Remove parent<->child consistency validation.
As noted now in update_build.go:

We used to validate consistency of timestamps and status between
parent and child. However with client-side protocols such as luciexe,
the parent and child steps may actually belong to separate processes
on the machine, and can race each other in `bbagent`.

Additionally, there's no way to guarantee that these two processes
would have a consistent monotonic clock state that's shared between
them (this is possible, but would take a fair amount of work) and
events such as Daylight Savings Time shifts could lead to up to an
hour of inconsistency between step timestamps.

R=ddoman, nodir, yiwzhang

Change-Id: I5891d87bb1925bad2616c71109e171b9c7292286