[prpc] Implement request message compression
Add Client.EnableRequestCompression. If true and the request body is
large enough, the compress it and add "Content-Encoding: gzip" header.
The toggle is needed, and must be off by default, because older
pRPC servers do not support compressed requests.

Note that http package automatically inserts "Accept-Encoding: gzip"
header, so it is challenging to disable response compression at the
pRPC.Client level. Thus the toggle is EnableRequestCompression, as
opposed to EnableCompression.

R=vadimsh@chromium.org

Bug: 1166257
Change-Id: Ic19f8c141baaa7de2d1529ce913a4626e6ed7ab5