[resultdb] Allow inclusions in BatchCreateInvocations
This change allows the included_invocations field of the invocations
in the batch to be used to include other invocations in the same batch.

This restriction (that both invocations are created in the same batch)
is required to simplify permissions checking.
If the caller is allowed to create invocations A and B, it follows that
it should be allowed to include one into the other.
Also, the restriction enables us to register the inclusion
unconditionally, i.e. without having to check that either invocation
exists (since we are creating both in the same transaction).

R=nodir,chanli

Bug: 1056007
Change-Id: Ia251aa618ab542305cc971992e54414b9d571517