[lucictx] Add some helper functions for Deadline.
Add some methods to Deadline to improve reading/setting the float64
timestamp.

Add CheckDeadlines to calculate if the current wall-clock says we're past
the soft or hard deadlines granted by our parent context (typically the
parent process). The whole timeout business is a bit wishy-washy due to
the reliance on non-synchronized clocks between processes, but most of
the time (ha) I think this should be good enough to detect if we're in
The Danger Zone (tm) of the grace period.

R=tandrii, yiwzhang

Bug: 1173222
Change-Id: Ib560057501f0a36e7ae4c5b88f8b30998183dbfa