[resultdb] Fix RecorderMock
gomock.Eq generally doesn't necessarily work for protobuf messages.
This is why ShouldResembleProto exists.

Also currently the request comparison happens outside of the Convey,
so the error message is not a part of it, and goconvey's text diffing
doesn't work.

Rewrite the recorder mock to avoid gomock.Eq and use ShouldResembleProto
instead.

Also fix some lint.

Bug: 1137513
Change-Id: Ia934f74dbfb47ba4374db3eccb9ea9f496d5aebf