[resultdb] Allow new invocations in FINALIZING state.
With this change an invocation can be created ready to be finalized,
that is, it will not be able to be updated and will not be able to
include any other invocations beyond those it includes at creation time.

The purpose of allowing this is to create invocations that thinly wrap
others, e.g. to use as aliases, and not having to call an RPC to
finalize them, instead relying on the call to finalize the wrapped
invocation.

R=chanli,nodir,iannucci,ddoman

Bug: 1056007
Change-Id: I006757f57da701b85a154116a36edf7efc1e8e49
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2689442
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>