[lucictx] Simplify and redocument AdjustDeadline.
This is now known as TrackDeadline, and is simplified to discourage
shrinking soft_deadline.

This also optimizes the implementation so that if the soft_deadline
doesn't change (i.e. `ctx` doesn't have a shorter Deadline attached),
then the goroutine won't preemptively send on the cleanup channel, but
instead will wait for the parent processes' signal. This should reduce
potential raciness because it means that in a stacked process situation,
the parent process will always be the one sending the signal, meaning
the parent process can directly record the fact that it's activiting
the timeout condition.

Concretely, this means that swarming would signal bbagent, bbagent
would signal the recipe engine, and bbagent would be able to report
the 'timeout' status detail to buildbucket.

R=tandrii, yiwzhang

Bug: 1173222
Change-Id: I3d9b9682db1593c31efe6f7bebdef2631cc3d5a5