[buildbucket] allows git commit to have both id and ref
py:validate_gitiles_commit() allows a commit to have both id and ref.
https://source.chromium.org/chromium/infra/infra/+/master:appengine/cr-buildbucket/validation.py;l=90-97

go:validateCommit() doesn't allow a commit to have ref, if id is set.
https://source.chromium.org/chromium/infra/infra/+/master:go/src/go.chromium.org/luci/buildbucket/appengine/rpc/common.go;l=215-216

It caused InvalidArgument errors in cr-buildbucket, and this CL patches
it to match with the behaviour of the python equivalent.

Bug:1110990
Change-Id: Ifcdc427999e2583523db99cd0b8347b2d966e064
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2692844
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>