[lucictx] Simplify and redocument AdjustDeadline.
This is now known as TrackSoftDeadline, and is simplified in two ways:
  * New API discourages shrinking soft_deadline.
  * New API no longer returns `cleanup` channel, instead makes
    the equivalent available in context to any part of the process
    which wants to use the time to clean up after itself.

This also optimizes the implementation so that if the soft_deadline
doesn't change (i.e. `ctx` doesn't have a shorter Deadline attached),
then the goroutine won't preemptively send on the cleanup channel, but
instead will wait for the parent processes' signal. This should reduce
potential raciness because it means that in a stacked process situation,
the parent process will always be the one sending the signal, meaning
the parent process can directly record the fact that it's activiting
the timeout condition.

Concretely, this means that swarming would signal bbagent, bbagent
would signal the recipe engine, and bbagent would be able to report
the 'timeout' status detail to buildbucket.

R=tandrii, yiwzhang

Bug: 1173222
Change-Id: I3d9b9682db1593c31efe6f7bebdef2631cc3d5a5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2678772
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>