[bbagent] Handle user subprocess retcode in a better way.
Previously if the subprocess exited with a non-zero exit code it would
be logged with a cryptic "host environment callback failed:" plus stack
trace.

Now we save the subprocess Wait error (if any) for after we've posted
the final status to buildbucket, at which point we attempt to extract the
exit code from the error.

R=tandrii, yiwzhang

Bug: 1173222
Change-Id: I1ef3a109f2a11db0327263aa3c362949c81df5b5