[resultui] provide invocation/build state at page level.
Updates:
1. Provide invocation-state and build-state at page level instead of via
stand alone provider elements.
2. Move state error handling and clean up to page level.
3. Flatten directory structure under ./src/context.
4. Update imports accordingly.

Benefits:
1. State related errors can be handled/emitted at page level.
2. Code complexity is reduced.

After the change, state (and associated data) is not cached when
switching to a new page. This has no effect in the current
implementation because there's no way to switch between pages. In the
future, we can cache data at service level if needed.

R=nodir@chromium.org, nqmtuan@google.com

Bug: 1114935
Change-Id: I2b10e165f8d985a442971e3160ca74bcd89d2bc4