[cv][lease] fix lease extension on transaction retries.
This hasn't failed *yet* with real Datastore tests,
but I noticed it while debugging another test failure.

The only way to catch with memory DS fake is by simulating DS retries,
which although possible, isn't trivial. It isn't worth the effort
in this concrete case: the worst case would be failure to extend
the lease, which CV code must be able to handle anyway.

R=qyearsley, yiwzhang

Change-Id: Ic9246bc2fd5da54985a86f82c2ba31884215e0c7