[cache] ignore cache policy error
Currently, it strictly respects cache policies.
Then, it often ends up with `no more space to free` if it violates min space policy.

It can be more tolerant to the cache policy violation as it may not require the configured min space.
It will fail if the disk doesn't have enough space in the end.

This change will ignore the cache policy violation error to let it continue process.

Change-Id: I3ac4a5721c3f23d6484c428ccb4156b92d249e6b