[lucicfg] Don't use `\x...` escapes in strings.
Recent starlark-go change [1] outlawed them. The proper fix would be
to start using `bytes` type for binary data everywhere, but it is a
huge change. So switch to UTF-8 code point for now as a surrogate.

Also bump lucicfg version to indicate we have picked up `bytes`
Starlark change (even if not fully incorporated yet).

[1]: https://chromium.googlesource.com/external/github.com/google/starlark-go.git/+/ebe61bd709bf23d7baddbb34e79084d7d156be04

R=iannucci@chromium.org

Change-Id: Ie3ae983ffd9f3914242d7068c5d80e4080d777dc