[datastore] Remove a flaky and invalid test
A flaky test which blocks CQ occasionally. This test was trying to check
if this goroutine (http://shortn/_40N31YGdtC) has exited after `ctx.cancel()`
was called. It assumed it would receive a stop err in itemCh after trying
to read at most twice. But the assumption is wrong. The goroutine will EVENTUALLY
exits but it cannot guarantee on which iteration if another thread keeps
reading itemCh. Besides, in real use cases, after `cancel()`, there won’t
be any thread to continue to read from the itemCh. So this unit test is
invalid.

Bug: 1142317
Change-Id: Ic590a50bb662a6d0bf8569e364173315e66fce5c