cv: Add skeleton of cv/internal/bq for BQ-row-sending code
This adds a bq package with general BQ row sending functionality.
The client should be created once and installed in context, for
consistency with the tree-status and gerrit clients.

A mock is provided which should allow us to avoid trying to actually
send rows in tests, but which should allow for verifying what rows
would have been sent.

Bug: 1173168
Change-Id: I5cd1df22a9cb00dbbaef07e627c60802f133b431