[buildbucket/bb] Add the ability to set/block experiments in `bb add`.
This also factors out 'well known' experiment strings to a single place
to help avoid typobugs, and documents the 'luci.use_realms' experiment
in the buildbucket protos.

Additionally this fixes `bb` to use the standard logging output
configuration so that log levels and timestamps are correctly
presented to aid in debugging and increase visibility of Warning, Error
log messages.

R=smut, vadimsh

Bug: 1093970,1134774
Change-Id: I60a44ab4b83f735ad97751b41f69532c8730c80b