[auth] Add AsSessionUser RPC authority kind.
It can be used to make RPC calls on user's behalf when using
encrypted tokens stored in a user session (or other method that
can provider fresh user tokens).

It is different from AsUser, because it depends on a particular
authentication method (unlike AsUser, which doesn't care how the
incoming request was authenticated). Note that AsUser is still
deprecated and will be removed once there are no usage left.

R=tandrii@chromium.org
CC=weiweilin@google.com
BUG=712506

Change-Id: Icd2150426ed7c382e520233b3a3387bda47053d3