[server] Refactor server/secrets package.
Introduce SecretManagerStore which will implement secret.Store using
Google Secret Manager. Implement RandomSecret(...) in it using the
same logic as we had before:
  * Load a root key from GSM.
  * Use it with DerivedStore to derive RandomSecret(...).
  * Periodically reload it.

Previously this logic was part of Module implementation, making it
hard to reuse. It is now in SecretManagerStore, available for reuse
when implementing StoredSecret(...) in the next CL.

Remove unnecessary public APIs: secrets.Source and its implementations.

Remove loading the root secret from a local file. It was unused and
it is always better to use GSM in production and phony devsecret://
in development.

Introduce "devsecret-text://<plain-text>" for secrets that are
already ASCII text (e.g. passwords). Keep "devsecret://<base64>"
as well, since it is used.

R=tandrii@chromium.org
BUG=712506

Change-Id: Ib77775da97c689bae135ed912fba4787bbc16fa4