[resultdb] add newUploadTask
newUploadTask() calls stat() to find the size of the artifact content,
and returns an uploadTask with the artifact information.

newUploadTask() is called inside sinkServer.ReportTestResults() and
sinkServer.ReportInvocationLevelArtifacts(). It means that
Sink will return an error to the client if Sink fails to obtain
the information of the given artifact files,
when the request was received.

Change-Id: Icfc85c2e4c7fbbb62ab1a8806c6c2dc4a734a429
Bug: 1149662
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2824182
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>