[encryptedcookies] Local fake mode for development.
This is conceptually similar to GAE dev appserver User API fake,
except it tries to emulate auth.Session API as well, so code that
uses auth.AsSessionUser can be tested locally as well.

It can produce real tokens in auth.Session only when the logged in
user matches the server's own account (which happens when the
developer launches the server process as self, and the goes to the
fake login page and logs in as self).

R=tandrii@chromium.org
BUG=712506

Change-Id: Iab0a09c1ef088837c02a4352d81bd9510d5b8a6c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2837613
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>