[resultdb] ignore artifact file error
In https://crrev.com/c/2824182, a change was made such that
the artifact file is checked when the request was received.

It turned out that blink web tests have been reporting artifacts
with non-existing files, but the errors were swallowed by
the dispatcher goroutines, as crbug.com/1124868 describes.

This patch introduces cfg.ExitOnFailures and update SinkServer
such that the artifact file error is ignored if it is set false.

Bug: 1124868
Change-Id: I9c3c889b57172ef1d7f714f8075196371032f008