[server] Make router.Middleware a slice.
It appears there's no real value in wrapping the slice in
a struct, but there's a big downside: being a struct,
router.Middleware is not "nillable". It forces a lot of
boilerplaty router.Middleware{} calls in a lot of places.

They'll be removed in the next CL.

R=tandrii@chromium.org, iannucci@chromium.org

Change-Id: Icaeb9be6125cc885dd7cfb0c63ed207bd7dbd4fb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2849905
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>