[resultdb] ignore artifact file error
In https://crrev.com/c/2824182, a change was made such that
the artifact file is checked when the request was received.

It turned out that blink web tests have been reporting artifacts
with non-existing files, but the errors were swallowed by
the dispatcher goroutines, as crbug.com/1124868 describes.

This patch updates SinkServer such that the artifact file error is
ignored.

Bug: 1124868
Change-Id: I9c3c889b57172ef1d7f714f8075196371032f008
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2855417
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>