[server] Add "go.chromium.org/luci/server/cron" module.
It "standardizes" how we register cron handlers, simplifying
the transition away from GAEv1 by:

- Implementing appropriate authentication e.g. using OpenID tokens
  when using Cloud Scheduler jobs.
- Providing a local UI for testing cron jobs as a replacement for
  the dev appserver UI.
- Automatically exposing monitoring metrics, making it possible to
  setup generic alerts.

It is meant as a replacement for gaemiddleware.RequireCron which is
not portable to GKE.

Switch `tokenserver` and `cipd` as a usage example. It is rather
trivial if existing handlers are already exposed via /internal/cron.

R=tandrii@chromium.org

Change-Id: I4567d33e39ac00979e7f7c09c7e5f5460bc4a0df