[logdog] Convert `logs` module to use server.Server.
This also switches Logdog to use encrypted cookies for login. This
change could not be made in isolation, since encrypted cookies
session implementation needs write access to the datastore, even
when just "validating" cookies (to be able to refresh the session
when tokens expire). But `logs` module has read only access.

Conversely, the current go111 cookies implementation doesn't work
with server.Server (being deprecated and all).

Note that using datastore to store sessions is safe, since it is
done exclusively from `logs` module, without any caching (for now).

Additionally, this is the first time server.Server is used for Flex.
It needed some minor adjustments. One complication is that there
appears to be no easy way to pass CLI flags to the binary, so we
use env vars instead and construct various ModuleOptions manually.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1204268

Change-Id: I52183dca31cfe967f915dd5494cc5d88a7f86862